refactor: naver/service/vo 패키지에서 config 패키지로 이동
- OAuth Client, Registration 환경 설정 정보를 담고 있는 객체이므로
- config 패키지로 이동시킴
