+++ src/main/java/com/ludo/study/studymatchingplatform/auth/naver/repository/InMemoryClientRegistrationAndProviderRepository.java
@@ -2,7 +2,7 @@ package com.ludo.study.studymatchingplatform.auth.naver.repository;
 
 import java.util.Map;
 
-import com.ludo.study.studymatchingplatform.auth.naver.service.vo.property.ClientRegistrationAndProvider;
+import com.ludo.study.studymatchingplatform.auth.config.property.ClientRegistrationAndProvider;
 import com.ludo.study.studymatchingplatform.user.domain.Social;
 
 import lombok.RequiredArgsConstructor;
