+++ src/main/java/com/ludo/study/studymatchingplatform/config/ClientRegistrationAndProviderRepositoryConfig.java
@@ -6,10 +6,10 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+import com.ludo.study.studymatchingplatform.auth.config.property.ClientRegistrationAndProvider;
+import com.ludo.study.studymatchingplatform.auth.config.property.ClientRegistrationAndProviderMapper;
+import com.ludo.study.studymatchingplatform.auth.config.property.OAuthProperties;
 import com.ludo.study.studymatchingplatform.auth.naver.repository.InMemoryClientRegistrationAndProviderRepository;
-import com.ludo.study.studymatchingplatform.auth.naver.service.vo.property.ClientRegistrationAndProvider;
-import com.ludo.study.studymatchingplatform.auth.naver.service.vo.property.ClientRegistrationAndProviderMapper;
-import com.ludo.study.studymatchingplatform.auth.naver.service.vo.property.OAuthProperties;
 
 import lombok.extern.slf4j.Slf4j;
